Adds nwb readable interface
