+++ allensdk/brain_observatory/behavior/data_objects/_base/_data_object_abc.py
@@ -23,20 +23,3 @@ class DataObject(abc.ABC):
     @property
     def value(self) -> Any:
         return self._value
-
-    @classmethod
-    @abc.abstractmethod
-    def from_nwb(cls, nwbfile: NWBFile) -> "DataObject":  # pragma: no cover
-        """Populate a DataObject from a pyNWB file object.
-
-        Parameters
-        ----------
-        nwbfile:
-            The file object (NWBFile) of a pynwb dataset file.
-
-        Returns
-        -------
-        DataObject:
-            An instantiated DataObject which has `name` and `value` properties
-        """
-        raise NotImplementedError()
