+++ allensdk/brain_observatory/behavior/data_objects/metadata/behavior_metadata/behavior_metadata.py
@@ -16,6 +16,9 @@ from allensdk.brain_observatory.behavior.data_objects._base\
     .readable_interfaces.json_readable_interface import \
     JsonReadableInterface
 from allensdk.brain_observatory.behavior.data_objects._base\
+    .readable_interfaces.nwb_readable_interface import \
+    NwbReadableInterface
+from allensdk.brain_observatory.behavior.data_objects._base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
 from allensdk.brain_observatory.behavior.data_objects._base.writable_interfaces\
@@ -186,6 +189,7 @@ def get_task_parameters(data: Dict) -> Dict:
 
 class BehaviorMetadata(DataObject, InternalReadableInterface,
                        JsonReadableInterface,
+                       NwbReadableInterface,
                        JsonWritableInterface,
                        NwbWritableInterface):
     """Container class for behavior metadata"""
