+++ allensdk/brain_observatory/behavior/data_objects/metadata/behavior_metadata/behavior_session_id.py
@@ -10,6 +10,9 @@ from allensdk.brain_observatory.behavior.data_objects._base.readable_interfaces\
     .lims_readable_interface import \
     LimsReadableInterface
 from allensdk.brain_observatory.behavior.data_objects._base\
+    .readable_interfaces.nwb_readable_interface import \
+    NwbReadableInterface
+from allensdk.brain_observatory.behavior.data_objects._base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
 from allensdk.internal.api import PostgresQueryMixin
@@ -22,6 +25,7 @@ def from_lims_cache_key(cls, db, ophys_experiment_id: int):
 
 class BehaviorSessionId(DataObject, LimsReadableInterface,
                         JsonReadableInterface,
+                        NwbReadableInterface,
                         JsonWritableInterface):
     def __init__(self, behavior_session_id: int):
         super().__init__(name="behavior_session_id", value=behavior_session_id)
