+++ allensdk/brain_observatory/behavior/data_objects/metadata/behavior_metadata/behavior_session_uuid.py
@@ -5,12 +5,16 @@ from pynwb import NWBFile
 
 from allensdk.brain_observatory.behavior.data_files import StimulusFile
 from allensdk.brain_observatory.behavior.data_objects import DataObject
+from allensdk.brain_observatory.behavior.data_objects._base\
+    .readable_interfaces.nwb_readable_interface import \
+    NwbReadableInterface
 from allensdk.brain_observatory.behavior.data_objects._base.readable_interfaces\
     .stimulus_file_readable_interface import \
     StimulusFileReadableInterface
 
 
-class BehaviorSessionUUID(DataObject, StimulusFileReadableInterface):
+class BehaviorSessionUUID(DataObject, StimulusFileReadableInterface,
+                          NwbReadableInterface):
     """the universally unique identifier (UUID)"""
     def __init__(self, behavior_session_uuid: Optional[uuid.UUID]):
         super().__init__(name="behavior_session_uuid", value=behavior_session_uuid)
