+++ allensdk/brain_observatory/behavior/data_objects/metadata/behavior_metadata/equipment_name.py
@@ -8,13 +8,16 @@ from allensdk.brain_observatory.behavior.data_objects._base\
     .readable_interfaces.lims_readable_interface import \
     LimsReadableInterface
 from allensdk.brain_observatory.behavior.data_objects._base\
+    .readable_interfaces.nwb_readable_interface import \
+    NwbReadableInterface
+from allensdk.brain_observatory.behavior.data_objects._base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
 from allensdk.internal.api import PostgresQueryMixin
 
 
 class EquipmentName(DataObject, JsonReadableInterface, LimsReadableInterface,
-                    JsonWritableInterface):
+                    NwbReadableInterface, JsonWritableInterface):
     """the name of the experimental rig."""
     def __init__(self, equipment_name: str):
         super().__init__(name="equipment_name", value=equipment_name)
