+++ allensdk/brain_observatory/behavior/data_objects/metadata/behavior_metadata/session_type.py
@@ -2,13 +2,17 @@ from pynwb import NWBFile
 
 from allensdk.brain_observatory.behavior.data_files import StimulusFile
 from allensdk.brain_observatory.behavior.data_objects import DataObject
+from allensdk.brain_observatory.behavior.data_objects._base\
+    .readable_interfaces.nwb_readable_interface import \
+    NwbReadableInterface
 from allensdk.brain_observatory.behavior.data_objects._base.readable_interfaces\
     .stimulus_file_readable_interface \
     import \
     StimulusFileReadableInterface
 
 
-class SessionType(DataObject, StimulusFileReadableInterface):
+class SessionType(DataObject, StimulusFileReadableInterface,
+                  NwbReadableInterface):
     """the stimulus set used"""
     def __init__(self, session_type: str):
         super().__init__(name="session_type", value=session_type)
