+++ allensdk/brain_observatory/behavior/data_objects/metadata/subject_metadata/age.py
@@ -12,13 +12,16 @@ from allensdk.brain_observatory.behavior.data_objects._base\
     .readable_interfaces.lims_readable_interface import \
     LimsReadableInterface
 from allensdk.brain_observatory.behavior.data_objects._base\
+    .readable_interfaces.nwb_readable_interface import \
+    NwbReadableInterface
+from allensdk.brain_observatory.behavior.data_objects._base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
 from allensdk.internal.api import PostgresQueryMixin
 
 
 class Age(DataObject, JsonReadableInterface, LimsReadableInterface,
-          JsonWritableInterface):
+          NwbReadableInterface, JsonWritableInterface):
     """Age of animal (in days)"""
     def __init__(self, age: int):
         super().__init__(name="age_in_days", value=age)
