+++ allensdk/brain_observatory/behavior/data_objects/metadata/subject_metadata/driver_line.py
@@ -10,6 +10,9 @@ from allensdk.brain_observatory.behavior.data_objects._base.readable_interfaces\
     .lims_readable_interface import \
     LimsReadableInterface
 from allensdk.brain_observatory.behavior.data_objects._base\
+    .readable_interfaces.nwb_readable_interface import \
+    NwbReadableInterface
+from allensdk.brain_observatory.behavior.data_objects._base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
 from allensdk.internal.api import PostgresQueryMixin, \
@@ -17,7 +20,7 @@ from allensdk.internal.api import PostgresQueryMixin, \
 
 
 class DriverLine(DataObject, LimsReadableInterface, JsonReadableInterface,
-                 JsonWritableInterface):
+                 NwbReadableInterface, JsonWritableInterface):
     """the genotype name(s) of the driver line(s)"""
     def __init__(self, driver_line: List[str]):
         super().__init__(name="driver_line", value=driver_line)
