+++ allensdk/brain_observatory/behavior/data_objects/metadata/subject_metadata/full_genotype.py
@@ -11,13 +11,16 @@ from allensdk.brain_observatory.behavior.data_objects._base.readable_interfaces\
     .lims_readable_interface import \
     LimsReadableInterface
 from allensdk.brain_observatory.behavior.data_objects._base\
+    .readable_interfaces.nwb_readable_interface import \
+    NwbReadableInterface
+from allensdk.brain_observatory.behavior.data_objects._base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
 from allensdk.internal.api import PostgresQueryMixin
 
 
 class FullGenotype(DataObject, LimsReadableInterface, JsonReadableInterface,
-                   JsonWritableInterface):
+                   NwbReadableInterface, JsonWritableInterface):
     """the name of the subject's genotype"""
     def __init__(self, full_genotype: str):
         super().__init__(name="full_genotype", value=full_genotype)
