+++ allensdk/brain_observatory/behavior/data_objects/metadata/subject_metadata/reporter_line.py
@@ -11,6 +11,9 @@ from allensdk.brain_observatory.behavior.data_objects._base.readable_interfaces\
     .lims_readable_interface import \
     LimsReadableInterface
 from allensdk.brain_observatory.behavior.data_objects._base\
+    .readable_interfaces.nwb_readable_interface import \
+    NwbReadableInterface
+from allensdk.brain_observatory.behavior.data_objects._base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
 from allensdk.internal.api import PostgresQueryMixin, \
@@ -18,7 +21,7 @@ from allensdk.internal.api import PostgresQueryMixin, \
 
 
 class ReporterLine(DataObject, LimsReadableInterface, JsonReadableInterface,
-                   JsonWritableInterface):
+                   NwbReadableInterface, JsonWritableInterface):
     """the genotype name(s) of the reporter line(s)"""
     def __init__(self, reporter_line: Optional[str]):
         super().__init__(name="reporter_line", value=reporter_line)
