+++ allensdk/brain_observatory/behavior/data_objects/metadata/subject_metadata/subject_metadata.py
@@ -8,6 +8,9 @@ from allensdk.brain_observatory.behavior.data_objects._base.readable_interfaces\
     .lims_readable_interface import \
     LimsReadableInterface
 from allensdk.brain_observatory.behavior.data_objects._base\
+    .readable_interfaces.nwb_readable_interface import \
+    NwbReadableInterface
+from allensdk.brain_observatory.behavior.data_objects._base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
 from allensdk.brain_observatory.behavior.data_objects._base.writable_interfaces\
@@ -36,7 +39,7 @@ from allensdk.brain_observatory.nwb import load_pynwb_extension
 from allensdk.internal.api import PostgresQueryMixin
 
 
-class SubjectMetadata(DataObject, LimsReadableInterface,
+class SubjectMetadata(DataObject, LimsReadableInterface, NwbReadableInterface,
                       NwbWritableInterface, JsonWritableInterface):
     """Subject metadata"""
     def __init__(self,
