+++ allensdk/brain_observatory/behavior/data_objects/running_speed/running_speed.py
@@ -17,6 +17,9 @@ from allensdk.brain_observatory.behavior.data_objects._base\
     .readable_interfaces.lims_readable_interface import \
     LimsReadableInterface
 from allensdk.brain_observatory.behavior.data_objects._base\
+    .readable_interfaces.nwb_readable_interface import \
+    NwbReadableInterface
+from allensdk.brain_observatory.behavior.data_objects._base\
     .writable_interfaces.json_writable_interface import \
     JsonWritableInterface
 from allensdk.brain_observatory.behavior.data_objects._base\
@@ -53,8 +56,9 @@ def from_lims_cache_key(
     )
 
 
-class RunningSpeed(DataObject, LimsReadableInterface, NwbWritableInterface,
-                   JsonReadableInterface, JsonWritableInterface):
+class RunningSpeed(DataObject, LimsReadableInterface, NwbReadableInterface,
+                   NwbWritableInterface, JsonReadableInterface,
+                   JsonWritableInterface):
     """A DataObject which contains properties and methods to load, process,
     and represent running speed data.
